refactor: refactored code base
