+++ src/lib/guides/index.ts
@@ -1,1 +1,4 @@
 export * from './getAllGuides';
+export * from './getDocumentsMenu';
+export * from './getFileContents';
+export * from './getPriorityWiseGuides';
